kickstart-learning-using-agents/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
uvloop==0.21.0
jinja2==3.1.4
python-multipart==0.0.17
python-dotenv==1.1.0
langgraph==0.3.31
langchain==0.3.23
langchain-community==0.3.21
langchain_google_genai==2.1.3
tavily-python==0.5.4